Clean up AuthorizationOptions.
