+++ packages/ai/src/ciba/auth0/polling-authorizer.ts
@@ -1,5 +1,5 @@
+import { AuthorizationOptions } from '../../authorizer'
 import { PollingCIBAAuthorizer, CIBAAuthorizerOptions } from '../polling-authorizer'
-import { AuthorizationError, AuthorizationOptions } from '../../errors/authorizationerror'
 
 interface Auth0CIBAAuthorizerOptions extends CIBAAuthorizerOptions {
   domain?: string;
